+++ bin/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports.buckets = require('./actions/buckets');
+
+module.exports.account = require('./actions/account');
+
+module.exports.keys = require('./actions/keys');
